+++ UtilitiesTests/DownloadTests.cs
@@ -5,9 +5,20 @@ public class DownloadTests
     [Fact]
     public void GetUriInformation()
     {
-        Uri uri = new(
-            "https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png"
-        );
-        _ = Download.GetContentInfo(uri, out long _, out DateTime _).Should().BeTrue();
+        using LoopbackServer server = new();
+
+        _ = Download.GetContentInfo(server.OkUri, out long size, out DateTime _).Should().BeTrue();
+        _ = size.Should().Be(LoopbackServer.Content.Length);
+    }
+
+    [Fact]
+    public void GetContentInfo_WithNotFoundUri_ShouldReturnFalse()
+    {
+        using LoopbackServer server = new();
+
+        _ = Download
+            .GetContentInfo(server.MissingUri, out long _, out DateTime _)
+            .Should()
+            .BeFalse();
     }
 }